refactor: cookie관련 로직 service에서 controller로 변경
