+++ src/main/java/com/prgrms/domain/user/UserController.java
@@ -4,10 +4,12 @@ import static com.prgrms.dto.UserDto.Request;
 
 import com.prgrms.dto.UserDto.Request.Login;
 import com.prgrms.dto.UserDto.Response;
+import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.net.URI;
 import java.nio.file.AccessDeniedException;
+import java.util.Arrays;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -29,13 +31,13 @@ public class UserController {
     public ResponseEntity<Response> getLoggedUserInfo(HttpServletRequest request)
         throws AccessDeniedException {
 
-        long loggedUserId = Long.parseLong(service.getCookieUserId(request).getValue());
+        long loggedUserId = Long.parseLong(getCookieUserId(request).getValue());
         Response userInfo = service.findUserById(loggedUserId);
 
         return ResponseEntity.ok(userInfo);
     }
 
-    @PostMapping("/new")
+    @PostMapping()
     public ResponseEntity<Void> userSignUp(@RequestBody Request userDto) {
 
         service.insertUser(userDto);
@@ -49,11 +51,13 @@ public class UserController {
     public ResponseEntity<Void> login(@RequestBody Login loginDto,
         HttpServletResponse servletResponse) {
 
-        Response response = service.login(loginDto, servletResponse);
+        Response response = service.login(loginDto);
+        setCookieUserId(response.getUserId(), servletResponse);
 
-        URI myPageUri = URI.create("api/users/me" + response.getUserId());
+        URI myPageUri = URI.create("api/users/me");
 
-        return ResponseEntity.created(myPageUri)
+        return ResponseEntity.ok()
+            .location(myPageUri)
             .build();
     }
 
@@ -61,9 +65,27 @@ public class UserController {
     public ResponseEntity<Void> logout(HttpServletResponse response,HttpServletRequest request)
         throws AccessDeniedException {
 
-        service.logout(request, response);
+        Cookie cookieUserId = getCookieUserId(request);
+        cookieUserId.setValue("");
+        response.addCookie(cookieUserId);
 
-        return ResponseEntity.ok().build();
+        return ResponseEntity.ok()
+            .build();
+    }
+
+    private void setCookieUserId(Long userId, HttpServletResponse response) {
+
+        Cookie userCookie = new Cookie("userId", String.valueOf(userId));
+        response.addCookie(userCookie);
+    }
+
+    private Cookie getCookieUserId(HttpServletRequest request) throws AccessDeniedException {
+
+        return Arrays.stream(request.getCookies())
+            .filter(i -> "userId".equals(i.getName()))
+            .filter(i -> !"".equals(i.getValue()))
+            .findAny()
+            .orElseThrow(() -> new AccessDeniedException("접근 권한이 없습니다"));
     }
 
 }
