+++ src/main/java/com/prgrms/domain/user/UserService.java
@@ -7,12 +7,7 @@ import com.prgrms.dto.UserDto.Response;
 import com.prgrms.exception.ErrorCode;
 import com.prgrms.exception.customException.EmailDuplicateException;
 import com.prgrms.exception.customException.UserNotFoundException;
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
-import java.nio.file.AccessDeniedException;
-import java.util.Arrays;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.validation.annotation.Validated;
@@ -22,9 +17,6 @@ import org.springframework.validation.annotation.Validated;
 public class UserService {
 
     private final UserRepository userRepository;
-
-    private final String COOKIE_NAME_USER_ID = "userId";
-    private final String EMPTY = "";
 
     public UserService(UserRepository userRepository) {
         this.userRepository = userRepository;
@@ -55,38 +47,11 @@ public class UserService {
     }
 
     @Transactional(readOnly = true)
-    public Response login(@Valid Login loginDto, HttpServletResponse servletResponse) {
+    public Response login(@Valid Login loginDto) {
 
-        Response response = userRepository.findUser(loginDto.email(), loginDto.password())
+        return userRepository.findUser(loginDto.email(), loginDto.password())
             .map(Response::new)
             .orElseThrow(() -> new UserNotFoundException(ErrorCode.USER_NOT_FOUND));
-
-        setCookieUserId(response.getUserId(), servletResponse);
-
-        return response;
-    }
-
-    public void logout(HttpServletRequest request, HttpServletResponse response)
-        throws AccessDeniedException {
-
-        Cookie cookieUserId = getCookieUserId(request);
-        cookieUserId.setValue(EMPTY);
-        response.addCookie(cookieUserId);
-    }
-
-    public Cookie getCookieUserId(HttpServletRequest request) throws AccessDeniedException {
-
-        return Arrays.stream(request.getCookies())
-            .filter(i -> COOKIE_NAME_USER_ID.equals(i.getName()))
-            .filter(i -> !EMPTY.equals(i.getValue()))
-            .findAny()
-            .orElseThrow(() -> new AccessDeniedException("접근 권한이 없습니다"));
-    }
-
-    private void setCookieUserId(Long userId, HttpServletResponse response) {
-
-        Cookie userCookie = new Cookie(COOKIE_NAME_USER_ID, String.valueOf(userId));
-        response.addCookie(userCookie);
     }
 
 }
